'solutionv3'

+++ app/shop.py
@@ -5,7 +5,7 @@ class Shop:
     def __init__(self,
                  name: str,
                  location: list[int],
-                 products: dict[str, int]
+                 products: dict[str, float]
                  ) -> None:
         self.name = name
         self.location = location
